pyproject.toml
[tool.poetry]
name = "elearning-platform-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-storages = "^1.13.2"
django-environ = "^0.10.0"
boto3 = "^1.26.138"
drf-spectacular = "^0.26.2"
psycopg2-binary = "^2.9.6"
dj-database-url = "^2.0.0"
pyjwt = "^2.7.0"
ruff = "^0.0.271"
djangorestframework-simplejwt = "^5.2.2"
coreapi = "^2.3.3"
pyyaml = "^6.0"
factory-boy = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
